Removing unneeded controller and adding logic back into TextController.
